Refactoring to custom Rspec matcher
This is a rather major refactor and was in fact too large for a single
commit. There is a LOT going on with this commit. So I apologize for
that.

However, you now see how to move towards a custom matcher. I'm still
not satisified as I beleive the ActiveFedora::PersistenceSpecHelper
should be a class instead of a module; Sharing knowledge of instance
variables across modules is a dangerous plan.

However, I have settled on the API for the persistence tests. Which
means I can grab the screen shots and proceed.